% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{as.data.frame.exclude}
\alias{as.data.frame.exclude}
\title{Convert an exclude object to a data frame}
\usage{
\method{as.data.frame}{exclude}(x, ...)
}
\arguments{
\item{x}{An object of class exclude}

\item{...}{Dummy. Only needed to satisfy the requirement from the generic as.data.frame function.}
}
\value{
A data.frame.
}
\description{
Convert an exclude object to a data frame
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
e <- get_exclude()
as.data.frame(e)
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
