% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{as_tibble.exclude}
\alias{as_tibble.exclude}
\title{Convert an exclude object to tibble}
\usage{
\method{as_tibble}{exclude}(x, ...)
}
\arguments{
\item{x}{An object of class exclude}

\item{...}{Dummy}
}
\value{
A Tibble.
}
\description{
Convert an exclude object to tibble
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
e <- get_exclude()
tibble::as_tibble(e)
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
