% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Record the exclusions performed on the dataframe.}
\usage{
exclude(data, what = "Exclusion", e_name = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{what}{Exclusion message.}

\item{e_name}{Name of the exclude object. By default the name of the current exclusion object is used.}
}
\value{
Returns the dataframe invisibly.
}
\description{
Record the exclusions performed on the dataframe.
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
\seealso{
\code{\link[=init_exclude]{init_exclude()}} to start tracking exclusions
}
