% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{get_exclude}
\alias{get_exclude}
\title{Get the exclude object with name e_name}
\usage{
get_exclude(e_name = NULL)
}
\arguments{
\item{e_name}{Name of the wanted exclude object. Default value is the name of the current exclude object.}
}
\value{
An exclude object.
}
\description{
Get the exclude object with name e_name
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
get_exclude()
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
