% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{init_exclude}
\alias{init_exclude}
\title{Initialize an exclude object}
\usage{
init_exclude(
  data,
  e_name = "default",
  statistics = function(data) {
     list(count = nrow(data))
 },
  what = "Original data"
)
}
\arguments{
\item{data}{A dataframe on which exclusions will be performed.}

\item{e_name}{Name of the exclude object.}

\item{statistics}{A named list of statistic functions.}

\item{what}{Name for the initial dataset.}
}
\value{
Returns the dataframe invisibly.
}
\description{
Initialize an exclude object
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
init_exclude(mtcars)
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
\seealso{
\code{\link[=exclude]{exclude()}} to log exclusions
}
