% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{plot_flow}
\alias{plot_flow}
\title{Visualize the exclusion flow as a string in the dot language}
\usage{
plot_flow(df)
}
\arguments{
\item{df}{A data frame in form returned by as_tibble or as.data.frame}
}
\value{
A character vector describing are graph in the dot language
}
\description{
The dot string can then be saved to a file with \code{cat(plot_flow(as_tibble(e)), file="example.dot")}
call, and the converted to png format using the following command:
\verb{dot example.dot -T png -o example.png}
}
\examples{
library(magrittr)
library(tibble)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
e <- get_exclude()
plot_flow(as_tibble(e))
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
