% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{print.exclude}
\alias{print.exclude}
\title{Print the exclude object}
\usage{
\method{print}{exclude}(x, ...)
}
\arguments{
\item{x}{Object of class exclude.}

\item{...}{Just to satisfy the requirement of the generic function.}
}
\value{
None
}
\description{
Print the exclude object
}
\examples{
library(magrittr)
old <- exclude:::pop("default") # Only needed in the help page. Keeps environment clean.
mtcars \%>\% 
  init_exclude() \%>\%
  dplyr::filter(gear != 3) \%>\%
  exclude()
e <- get_exclude()
print(e)
invisible(exclude:::push(old, "default"))  # Only needed in the help page. Restore environment.
}
