% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude-package.R
\docType{package}
\name{exclude-package}
\alias{exclude-package}
\title{exclude: Keeps track of data exclusions}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jttoivon/exclude}
  \item Report bugs at \url{https://github.com/jttoivon/exclude/issues}
}

}
\author{
\strong{Maintainer}: Jarkko Toivonen \email{jarkko.toivonen.19@gmail.com} (\href{https://orcid.org/0000-0002-6843-5831}{ORCID})

}
\keyword{internal}
